// src/intl.d.ts
export declare const commonMessages: {
    availability: {
        defaultMessage: string;
        id: string;
    };
    catalog: {
        defaultMessage: string;
        id: string;
    };
    chooseFile: {
        defaultMessage: string;
        id: string;
        description: string;
    };
    channel: {
        defaultMessage: string;
        id: string;
    };
    customApps: {
        defaultMessage: string;
        id: string;
    };
    dashboard: {
        defaultMessage: string;
        id: string;
    };
    demo: {
        defaultMessage: string;
        id: string;
        description: string;
    };
    description: {
        defaultMessage: string;
        id: string;
    };
    descriptionOptional: {
        defaultMessage: string;
        id: string;
    };
    discounts: {
        defaultMessage: string;
        id: string;
    };
    drafts: {
        defaultMessage: string;
        id: string;
    };
    email: {
        defaultMessage: string;
        id: string;
    };
    endDate: {
        defaultMessage: string;
        id: string;
    };
    endHour: {
        defaultMessage: string;
        id: string;
    };
    error: {
        defaultMessage: string;
        id: string;
    };
    firstName: {
        defaultMessage: string;
        id: string;
    };
    generalInformations: {
        defaultMessage: string;
        id: string;
    };
    insufficientPermissions: {
        defaultMessage: string;
        id: string;
    };
    lastName: {
        defaultMessage: string;
        id: string;
    };
    limitReached: {
        defaultMessage: string;
        id: string;
    };
    no: {
        defaultMessage: string;
        id: string;
    };
    optionalField: {
        defaultMessage: string;
        id: string;
        description: string;
    };
    properties: {
        defaultMessage: string;
        id: string;
    };
    readOnly: {
        defaultMessage: string;
        id: string;
    };
    requiredField: {
        defaultMessage: string;
        id: string;
    };
    savedChanges: {
        defaultMessage: string;
        id: string;
    };
    sessionExpired: {
        defaultMessage: string;
        id: string;
    };
    somethingWentWrong: {
        defaultMessage: string;
        id: string;
    };
    startDate: {
        defaultMessage: string;
        id: string;
    };
    startHour: {
        defaultMessage: string;
        id: string;
    };
    status: {
        defaultMessage: string;
        id: string;
    };
    summary: {
        defaultMessage: string;
        id: string;
    };
    translationAttributes: {
        defaultMessage: string;
        id: string;
    };
    unauthorizedDashboardAccess: {
        defaultMessage: string;
        id: string;
    };
    uploadImage: {
        defaultMessage: string;
        id: string;
        description: string;
    };
    yes: {
        defaultMessage: string;
        id: string;
    };
    date: {
        defaultMessage: string;
        id: string;
    };
    time: {
        defaultMessage: string;
        id: string;
        description: string;
    };
    welcome: {
        defaultMessage: string;
        id: string;
    };
};